Rename HornsAppBottomSheetFragment as ItemBottomSheetFragment
