+++ app/src/main/res/values/strings.xml
@@ -19,8 +19,8 @@
     <string name="error_default">Ha ocurrido un error, revisa tu conexión a internet y vuelve a intentarlo.</string>
     <string name="error_no_items">Por ahora no tenemos elementos para esta categoría.</string>
     <string name="error_no_favorite_yet">Aun no haz seleccionado ningun concierto como favorito.</string>
-    <string name="error_app_not_found">No tienes la aplicación instalada</string>
-    <string name="error_no_tickets">No esta activa la venta de entradas</string>
+    <string name="error_app_not_found">No tienes la aplicación instalada.</string>
+    <string name="error_no_tickets">No esta activa la venta de entradas.</string>
 
     <!-- Packages -->
     <string name="facebook_package">com.facebook.katana</string>
